Introduced InfusXmlService
