+++ src/MetadataProcessor/MetadataProcessorEngine.cs
@@ -1,10 +1,8 @@
 using Kurmann.Videoschnitt.MetadataProcessor.Services;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Kurmann.Videoschnitt.MetadataProcessor.Entities;
 using Kurmann.Videoschnitt.MetadataProcessor.Entities.SupportedMediaTypes;
 using CSharpFunctionalExtensions;
-using System.Xml.Linq;
 
 namespace Kurmann.Videoschnitt.MetadataProcessor
 {
@@ -19,11 +17,12 @@ namespace Kurmann.Videoschnitt.MetadataProcessor
         private readonly FFmpegMetadataService _ffmpegMetadataService;
         private readonly MediaTypeDetectorService _mediaTypeDetectorService;
         private readonly MediaSetVariantService _mediaSetVariantService;
+        private readonly InfuseXmlService _infuseXmlService;
 
         public MetadataProcessorEngine(IOptions<MetadataProcessorSettings> settings, ILogger<MetadataProcessorEngine> logger,
             MediaFileListenerService mediaFileListenerService, MetadataProcessingService metadataProcessingService, 
             FFmpegMetadataService ffmpegMetadataService, MediaTypeDetectorService mediaTypeDetectorService,
-            MediaSetVariantService mediaSetVariantService)
+            MediaSetVariantService mediaSetVariantService, InfuseXmlService infuseXmlService)
         {
             _settings = settings.Value;
             _mediaFileListenerService = mediaFileListenerService;
@@ -31,6 +30,7 @@ namespace Kurmann.Videoschnitt.MetadataProcessor
             _ffmpegMetadataService = ffmpegMetadataService;
             _mediaTypeDetectorService = mediaTypeDetectorService;
             _mediaSetVariantService = mediaSetVariantService;
+            _infuseXmlService = infuseXmlService;
         }
 
         public async Task<Result> Start(IProgress<string> progress)
@@ -82,7 +82,7 @@ namespace Kurmann.Videoschnitt.MetadataProcessor
                     // Prüfe ob eine QuickTime-Movie-Variante gefunden wurde
                     if (quickTimeMovieVariantResult.Value.HasNoValue)
                     {
-                        var readMpeg4MetadataResult = await ReadMetadataFromMpeg4Video(mpeg4Video, progress);
+                        var readMpeg4MetadataResult = await _infuseXmlService.ReadMetadataFromMpeg4Video(mpeg4Video, progress);
                         if (readMpeg4MetadataResult.IsFailure)
                         {
                             progress.Report(readMpeg4MetadataResult.Error);
@@ -98,7 +98,7 @@ namespace Kurmann.Videoschnitt.MetadataProcessor
                     progress.Report($"QuickTime-Movie-Variante für Mpeg4-Video {mpeg4Video.FileInfo.Name}: {quickTimeMovieVariantResult.Value.Value.FileInfo.Name}");
 
                     // Extrahiere Metadaten aus der QuickTime-Movie-Variante
-                    var readQuickTimeMetadataResult = await ReadMetdataFromQuickTimeMovie(quickTimeMovieVariantResult.Value.Value, progress);
+                    var readQuickTimeMetadataResult = await _infuseXmlService.ReadMetdataFromQuickTimeMovie(quickTimeMovieVariantResult.Value.Value, progress);
                     if (readQuickTimeMetadataResult.IsFailure)
                     {
                         progress.Report(readQuickTimeMetadataResult.Error);
@@ -114,55 +114,7 @@ namespace Kurmann.Videoschnitt.MetadataProcessor
             return Result.Success();
         }
 
-        private async Task<Result<XDocument>> ReadMetdataFromQuickTimeMovie(QuickTimeMovie quickTimeMovie, IProgress<string> progress)
-        {
-            progress.Report($"Extrahiere Metadaten aus QuickTime-Movie {quickTimeMovie.FileInfo.Name}");
-            var metadataResult = await _ffmpegMetadataService.GetFFmpegMetadataAsync(quickTimeMovie.FileInfo.FullName);
-            if (metadataResult.IsFailure)
-            {
-                return Result.Failure<XDocument>(($"Fehler beim Extrahieren der Metadaten aus QuickTime-Movie {quickTimeMovie.FileInfo.Name}: {metadataResult.Error}"));
-            }
 
-            // Parse die FFMpeg-Metadaten
-            var ffmpegMetadata = FFmpegMetadata.Create(metadataResult.Value);
-            if (ffmpegMetadata.IsFailure)
-            {
-                return Result.Failure<XDocument>($"Fehler beim Parsen der extrahierten Metadaten aus QuickTime-Movie {quickTimeMovie.FileInfo.Name}: {ffmpegMetadata.Error}");
-            }
-
-            // Informiere über die extrahierten Metadaten
-            progress.Report($"Extrahierte Metadaten aus QuickTime-Movie {quickTimeMovie.FileInfo.Name}: {ffmpegMetadata.Value}");
-
-            // Erstelle ein Infuse-XML-Objekt aus den Metadaten
-            var infuseXml = ffmpegMetadata.Value.ToInfuseXml();
-            return infuseXml;
-        }
-
-        private async Task<Result<XDocument>> ReadMetadataFromMpeg4Video(Mpeg4Video mpeg4Video, IProgress<string> progress)
-        {
-            progress.Report($"Extrahiere Metadaten aus Mpeg4-Video {mpeg4Video.FileInfo.Name}");
-            var metadataResult = await _ffmpegMetadataService.GetFFmpegMetadataAsync(mpeg4Video.FileInfo.FullName);
-            if (metadataResult.IsFailure)
-            {
-                return Result.Failure<XDocument>($"Fehler beim Extrahieren der Metadaten aus Mpeg4-Video {mpeg4Video.FileInfo.Name}: {metadataResult.Error}");
-            }
-
-            // Parse die FFMpeg-Metadaten
-            var ffmpegMetadata = FFmpegMetadata.Create(metadataResult.Value);
-            if (ffmpegMetadata.IsFailure)
-            {
-                return Result.Failure<XDocument>($"Fehler beim Parsen der extrahierten Metadaten aus Mpeg4-Video {mpeg4Video.FileInfo.Name}: {ffmpegMetadata.Error}");
-            }
-
-            // Informiere über die extrahierten Metadaten
-            progress.Report($"Extrahierte Metadaten aus Mpeg4-Video {mpeg4Video.FileInfo.Name}: {ffmpegMetadata.Value}");
-
-            // Erstelle ein Infuse-XML-Objekt aus den Metadaten
-            var infuseXml = ffmpegMetadata.Value.ToInfuseXml();
-            progress.Report($"Infuse-XML aus Metadaten von Mpeg4-Video {mpeg4Video.FileInfo.Name}: {infuseXml}");
-
-            return infuseXml;
-        }
     
     }
 }
